Fixed window icons not working on non-Windows machines.
Icons of type *.ico do not appear to work on non-Windows machines.
This patch provides an icon in png format and modifies main.py to
use the appropriate icon.

Tested on Debian GNU/Linux 12 (bookworm) KDE where it works
perfectly - ie the window icon and the icon on the taskbar are shown
correctly.

However when using Linux Mint 21.3 (Virginia) the window icon is not
shown (just a white space) but the taskbar icon is shown correctly.